third_party/blink/web_tests/http/tests/misc/extract-http-content-language-multiple-expected.txt

Test for bug 97929: Extract HTTP Content-Language header.

==> Value extracted singly from HTTP "Content-Language" header...
PASS window.getComputedStyle(document.getElementById("x")).webkitLocale is '"fr"'
==> Value set by div "lang" tag...
PASS window.getComputedStyle(document.getElementById("y")).webkitLocale is '"ar"'
==> All done...
PASS successfullyParsed is true

TEST COMPLETE
